refactor to use latest `vas` and `http-routes`
