+++ dex/service.js
@@ -4,5 +4,6 @@ module.exports = [
   require('dex/user/service'),
   require('dex/profiles/service'),
   require('dex/relationshipTypes/service'),
-  require('dex/relationships/service')
+  require('dex/relationships/service'),
+  require('dex/static/service')
 ]
