+++ package.json
@@ -53,6 +53,7 @@
     "bcrypt": "^0.8.7",
     "evalify": "^2.0.0",
     "get-form-data": "^1.2.5",
+    "http-routes": "github:ahdinosaur/http-routes#v1.0.0",
     "inu": "^3.1.3",
     "inu-log": "^1.0.2",
     "inux": "github:ahdinosaur/inux#navigate",
@@ -75,7 +76,7 @@
     "ticket-auth": "0.0.1",
     "tiny-route": "^2.1.2",
     "uglifyify": "^3.0.1",
-    "vas": "github:ahdinosaur/vas#all-the-betters",
+    "vas": "github:ahdinosaur/vas#more-better-server",
     "wzrd": "github:ahdinosaur/wzrd#patch-1",
     "xtend": "^4.0.1"
   }
